+++ NEWS.md
@@ -3,10 +3,14 @@
 * Creates and manipulates data frames using the new class `'genalex'`.  The `"genetic.data.format"` attribute present in earlier versions of this package is removed.
 * Extensive documentation updates reflecting use of class `'genalex'`
 * Several functions have been deprecated and replaced with generics recognising class `'genalex'`: `printGenalexGenotype` is replaced by `printGenotype`, `computeGenalexColumns` is replaced by `getLocusColumns`, `reorderGenalexLoci` is replaced by `reorderLoci`, `getGenalexLocus` is replaced by `getLocus`, `putGenalexLocus` is replaced by `replaceLocus`, `dropGenalexLoci` is replaced by `dropLocus`, and `reduceGenalexPloidy` is replaced by `reducePloidy`
+* Extended `is.genalex` to have shallow (default) or deep (`force = TRUE`) checking of class `'genalex'` consistency.  With deep checking, attributes are inferred from the structure of the data frame and compared to the actual attributes.  Checking of string attributes (`dataset.title`, `data.file.name`) and extra data columns may skipped with options.
 * New `as.genalex` generic with `as.genalex.genalex` and `as.genalex.data.frame`.  The latter will convert a pre-class `'genalex'` data frame to class `'genalex'`, and is used wherever a data frame is returned so conversion from old to new format happens automatically.  It will also convert a suitably-formatted data frame to class `'genalex'` after determining suitable attributes
 * There is a `summary` method for class `'genalex'` which prints a few lines describing the data set, followed by a summary of the data frame contents and a summary of the extra columns, if present
 * New `as.data.frame` method to convert class `'genalex'` to class `'data.frame'`, optionally clearing all class `'genalex'`-specific attributes
-* new `genalex` function to create a class `'genalex'` object from constituent data
+* New `genalex` function to create a class `'genalex'` object from constituent data
+* New `rbind` method to combine samples from multiple class `'genalex'` objects
+* New `cbind` method to combine loci and extra data columns from multiple class `'genalex'` objects
+* New `extra` and `extra<-` methods to get and set the extra columns of a class `'genalex'` object.  The latter coerces its `value` object to a data frame, if it is not already, does a bit of error checking, and sets the row names to be the sample names of the object.
 * `readGenalex` now has `...` in its argument list, for additional arguments to be passed to `scan()` while reading data
 * Added `na.strings` argument to `readGenalex`, with defaults that reflect GenAlEx's own expectations for missing data
 * Added `na` and `na.character` arguments to `writeGenalex`, with values to use for NA in genotype/numeric/logical columns and character columns, respectively